Adds changes to generate filename and name for directories
The create action of the Directory resource accepts a filename and a name.
If one is not given the other is generated.
When filename is missing is set to a lowercase, underscored version of the name.
When name is missing it is set to a humanized version of the filename.