Adds Info page.
This adds a LiveView InfoLive at the root / of the web app that shows
information about the xeno installation.

Now that is only the path to the directory containing the notes.
In the future there might be more.

This should be replaced with a proper start page when there is
functionality.