Working Create/Edit forms for Notes
This is all still a bit rudimentary.
The create form updates fields when the NoteType is changed
and has a field to select the directory.

The edit form does not allow changing the directory, filename,
or NoteType right now.

Later I think there will be a "move note" functionality and
a way to change the NoteType.
But that will probably have to way until the NoteType,
data validation, and note rendering functionalities have
progressed.